@@ -20,25 +20,27 @@
     "pigeon-maps": "^0.17.0",
     "pigeon-marker": "^0.3.4",
     "pigeon-overlay": "^0.2.3",
-    "react": "^16.13.0",
+    "react": "^18.3.1",
     "react-bootstrap": "^1.0.0-beta.14",
     "react-chartjs": "^1.2.0",
     "react-chartjs-2": "^2.8.0",
-    "react-dom": "^16.8.6",
+    "react-dom": "^18.3.1",
     "react-grid-gallery": "^0.5.5",
     "react-image-gallery": "^1.0.7",
     "react-loader-spinner": "^3.1.4",
-    "react-redux": "^7.1.0",
-    "react-redux-toastr": "^7.5.1",
+    "react-redux": "7",
+    "react-redux-toastr": "^8.0.0",
     "react-router": "^3.2.0",
-    "react-router-dom": "^5.0.0",
-    "react-scripts": "2.1.8",
+    "react-router-dom": "^6.26.0",
+    "react-scripts": "^5.0.1",
+    "react-toastify": "^10.0.5",
     "react-widgets": "^4.4.11",
     "react-widgets-moment": "^4.0.27",
     "redux": "^4.0.4",
     "redux-form": "^8.2.4",
     "redux-form-input-masks": "^2.0.1",
     "redux-thunk": "^2.3.0",
+    "use-sync-external-store": "^1.2.2",
     "yarn": "^1.21.1"
   },
   "scripts": {
